Update to General Climate Analysis
Removed TOBS (Temperature at Observation), not used since it is unlikely to be found in datasets. Removed dependency on TMAX and TMIN (Max and Minimum Temperature), can still run the analysis, but can ignore it and analyze precipitation only. Optimal for looking at snowfall and rainfall independently. In addition to looking at Rainfall and Snowfall extremes, Seasonality index and annual precipitation totals/types averages are used to smooth out variation. These averages are 3, 5, and 9 years. With the median being the central year of the average. This is why all averages were chosen as odd years. Snowfall is set to analyze 12 month periods starting in July, this ensures that all snowfall is accounted for between fall, winter, and spring. Annual temperature values are still based upon the 12 month period for snowfall. Needs to modified to do both still.